python-backend/requirements.txt
fastapi
uvicorn
openai
SpeechRecognition
gtts
pydub
python-multipart
python-dotenv
